WIP: Fix emit for IsLiteral static fields (errant code actually came from similar code in the DLR)

+++ Clojure/Clojure/CljCompiler/Ast/StaticFieldExpr.cs
@@ -137,16 +137,16 @@ namespace clojure.lang.CljCompiler.Ast
                             ilg.Emit(OpCodes.Ldc_I4_0);
                         }
                         break;
-                    case TypeCode.Byte: ilg.Emit(OpCodes.Ldc_I4, (byte)value); break;
-                    case TypeCode.Char: ilg.Emit(OpCodes.Ldc_I4, (char)value); break;
+                    case TypeCode.Byte: ilg.Emit(OpCodes.Ldc_I4, (int)(byte)value); break;
+                    case TypeCode.Char: ilg.Emit(OpCodes.Ldc_I4, (int)(char)value); break;
                     case TypeCode.Double: ilg.Emit(OpCodes.Ldc_R8, (double)value); break;
-                    case TypeCode.Int16: ilg.Emit(OpCodes.Ldc_I4, (short)value); break;
+                    case TypeCode.Int16: ilg.Emit(OpCodes.Ldc_I4, (int)(short)value); break;
                     case TypeCode.Int32: ilg.Emit(OpCodes.Ldc_I4, (int)value); break;
                     case TypeCode.Int64: ilg.Emit(OpCodes.Ldc_I8, (long)value); break;
-                    case TypeCode.SByte: ilg.Emit(OpCodes.Ldc_I4, (sbyte)value); break;
+                    case TypeCode.SByte: ilg.Emit(OpCodes.Ldc_I4, (int)(sbyte)value); break;
                     case TypeCode.Single: ilg.Emit(OpCodes.Ldc_R4, (float)value); break;
                     case TypeCode.String: ilg.Emit(OpCodes.Ldstr, (string)value); break;
-                    case TypeCode.UInt16: ilg.Emit(OpCodes.Ldc_I4, (ushort)value); break;
+                    case TypeCode.UInt16: ilg.Emit(OpCodes.Ldc_I4, (int)(ushort)value); break;
                     case TypeCode.UInt32: ilg.Emit(OpCodes.Ldc_I4, (uint)value); break;
                     case TypeCode.UInt64: ilg.Emit(OpCodes.Ldc_I8, (ulong)value); break;
                 }
